requirements.txt
torch
torchvision
timm
segmentation-models-pytorch
numpy
Pillow
matplotlib